Rename file according to coding standards
