+++ packages/api/cms-api/src/user-permissions/user-permissions.constants.ts
